feat : `MenuProductRequests` 제거 및 `Menu` DTO 클래스 수정
