+++ src/main/java/kitchenpos/dto/menu/MenuGroupRequest.java
@@ -18,6 +18,6 @@ public class MenuGroupRequest {
     }
 
     public MenuGroup toMenuGroup() {
-        return new MenuGroup(name);
+        return MenuGroup.of(name);
     }
 }
