+++ src/main/java/kitchenpos/dto/product/ProductRequest.java
@@ -25,6 +25,6 @@ public class ProductRequest {
     }
 
     public Product toProduct() {
-        return new Product(name, price);
+        return Product.of(name, price.intValue());
     }
 }
